Adding integration tests for GETREF and SLICE over JSON.
This commit also includes a change to IntraService. The SET op typically looks
like,

{
  "type": "SET",
  "op": {
    "rowkey": "row_key1",
    "name": "column1",
    "value": "value1"
  }
}

but when using GETREF it looks like,

{
  "type": "SET",
  "op": {
    "rowkey": "getref_test",
    "name": "col3",
    "value": {
      "type": "GETREF",
      "op": {
        "resultref": 4,
        "wanted": "value"
      }
    }
  }
}

In this case the value attribute references a nested GETREF op. With the
default JSON deserialization, the nested GETREF is getted deserialized as a
Map instead of a IntraOp. I have added logic in InstraService to handle the
nested map.