refactoring to replace GETREF IntraOp with a map
This change is a result of the discussion in
https://github.com/zznate/intravert-ug/pull/80. The GETREF IntraOp has been
removed with this commit. Its intended use was as the value of a SET IntraOp.
I already added logic in IntraService to handle a nested map in a SET command
since we do not have Jackson configured to deserialize a nested IntraOp.
IntraServiceITest has been refactored to use a nested map where GETREF was
previously used.