refactoring groovy filter support
The client can now provide a groovy filter that is either a closure or is a
direct implementation of Filter. The former can be as simple as,

// simple closure filter
{ row -> if (row['value'].toInteger() > 21) return row else return null }

and the latter might look like,

// filter that directly implements Filter
class Over21 implements Filter {
  Map filter(Map row) {
    if (row['value'].toInteger() > 21) {
      return row
    } else {
      return null
    }
  }
}

GroovyFilterFactory encapsulate the logic for creating a groovy filter. If the
client script returns a closure, then the factory will wrap it in a
GroovyFilter. If the script returns an instance of Filter than that object
itself is returned. If the script returns neither a closure nor a filter, then
the factory throws an exception. In subsequent commits I will provide similar
logic for other languages.