adding initial support for javascript filters
This is a bit different from how groovy is supported. For a JavaScript filter
you have to specify a single function as follows,

function over21(row] { if (row['value'] > 21) return row else return null; }

With the groovy implementation changes I made in 47737c4a4 you can define an
entire script, and the return value is assumed to be the filter. The JavaScript
version is different because the ECMA spec considers it illegal to have a
return statement outside of a function. But similar to a groovy script, the last
line of a javascript script is the return value of the script; so, it might be
possible to make behavior more consistent. For now though I am inclined to
leave it as is until I see how things look for other languages.

The other thing to point out with this commit is some of the work done in
JavaScriptFilter. A java.util.Map passed into a script is not treated as native
javascript object; so I first convert the row method argument into a
javascript object that is then passed to the function. JavaScript objects do
however implement Map so there is no conversion required for the return value
from the script function.

Lastly, integration tests have been updated as well.