initial support for processing a request through the event bus
When IntraHandlerJson gets called it sends a message on the event bus to the
address "request.json". The handler registered at that address sends a separate
message on the bus for each operation. When all of the operations have
completed we call the reply handler provided with the message sent to the
request.json address. That reply handler ends the http response.

So far I have this working with an initial test in RawJsonITest. Still several
things to be worked out like error handling and most importantly implementing
handlers for each of the defined operations.