adding set keyspace and create column family handlers
There is another change with this commit to point out. In the sync version of
these operations the set keyspace command sets the keyspace in a IntraState
object that lives for the duration of the request. I haven't yet bothered to
serialize an IntraState object, but put I have added a JsonObject named state
that is included in the message sent to each operation handler.

One drawback right now is that the operation handler has to be sure to include
the state in its response message. I will look to remove that requirement in a
future commit.