updating CQL support and test coverage
Adding a more comprehensive CQL test in RawJsonITest that creates a table,
inserts data, and then queries for it. The query results getting returned
looked like,

{
  "name":"username",
  "value":"jsmith"
},
{
  "name":"age",
  "value":"\u0000\u0000\u0000\u001E"
},
{
  "name":"fname",
  "value":"John"
},
{
  "name":"lname",
  "value":"Smith"
}

There are two problems with this. The first is that the value for the age
column was not properly converted. The second issue is that we are not
tranposing the results as you would expect with other CQL clients like cqlsh.
The above results in cqlsh would look something like,

username | fname | lname | age
------------------------------
jsmith     John    Smith   30

This commit returns the transposed results that look like,

{
  "lname":"Smith",
  "username":"jsmith",
  "age":30,
  "fname":"John"
}

Instead of returning 3 JSON objects, we return just the one. And also note that
the type conversion for the age column is handled. I really think this is the
way to go for the CQL query operation. If the client does not work the
transposed results then he can use a GET or SLICE operation.