make HandlerUtils.readCf create and return a JsonArray of columns
While I was looking at something else, I noticed that the callers of
HandlerUtils.readCf both take the list of maps that gets populated, iterates
through the list, converts each map to a JsonObject, and then adds that object
to a JsonArray that is sent back in the response to the reply handler.

This commit removes that duplicate work and also eliminates the extra work of
looping over all of the columns again.