first pass at support for filter handlers
When a request is created, a new FilterHandler is registered on the event bus.
A FilterHandler wraps a Filter object. The operation response is sent back on
the event bus after the filter has been applied to all of the rows.

This commit adds filter support to SliceHandler. I will add support in GetHandler
in my next commit. Then I need to see about minimizing some of the json object
creation/mapping overhead.