guard against NPEs
My previous changes to the readCf() method introduced some regressions around
the logic that determines how to calculate the column name and column value. I
needed to check if there is meta data for the column name and value. If and
only if there is meta data, the component value is calculated with a type
conversion; otherwise, we just take the raw byte buffer. Note that the
ByteBuffer cannot be stored as is in JsonObject. We need to get the bytes
which are in turn encoded. We might want to see about providing an alternate
implementation that will be more efficient, if that is possible.