Send a 400 response back to the client if there is a mapping exception
This is some initial error handling support which I am sure is likely to change
one all of the operations are ported to handlers. Prior to this though we were
simply swallowing the exception. The client would hang indefinitely waiting for
a response, making debugging difficult. With this we are propagating the stack
trace back to the client which should really help with debugging errors.